% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce_analysis.R
\name{compare_dirs}
\alias{compare_dirs}
\title{compare directories}
\usage{
compare_dirs(to = getwd())
}
\arguments{
\item{to}{Character scalar, file path, where to copy analysis structure
(default = current working directory).}
}
\description{
compare analysis directories in source and target directory
}
