% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce_analysis.R
\name{compute_hashes}
\alias{compute_hashes}
\title{compute hashes}
\usage{
compute_hashes(dir)
}
\arguments{
\item{dir}{Character, file path.}
}
\description{
compute md5 hashes for all files in specified directory.
}
