% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce_analysis.R
\name{copy_analysis}
\alias{copy_analysis}
\title{Copy analysis directory structure}
\usage{
copy_analysis(to = getwd(), drake = TRUE, overwrite = FALSE)
}
\arguments{
\item{to}{Character scalar, file path, where to copy the workflow's files to
(default = current working directory).}

\item{drake}{Logical, whether to copy \code{.drake} directory (default = \code{TRUE}).}

\item{overwrite}{Logical, whether to overwrite the analysis directory structure (default = \code{FALSE})}
}
\description{
Copy the analysis directory structure of a packaged \code{drake} workflow to the
selected directory.
}
\note{
This function will copy the analysis directory structure supplied with the package
and the \code{.drake/} directory into the user's working directory (or the directory supplied
to the \code{to} parameter).
}
