% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_drakepkg_files.R
\name{copy_drakepkg_files}
\alias{copy_drakepkg_files}
\title{Save the Source Code Files of the \code{drakepkg} Workflow}
\usage{
copy_drakepkg_files(to = getwd(), directories = c("extdata", "intdata",
  "documents"), overwrite_dirs = c(extdata = TRUE, intdata = TRUE, documents =
  TRUE), package = "drakepkg")
}
\arguments{
\item{to}{Character scalar, file path, where to write the folders containing the workflow's files.}

\item{directories}{Character vector, names of the directories to copy from the source package to the target location}

\item{overwrite_dirs}{Logical vector (named), A vector of logical values with names that exactly match the \code{directories} values}

\item{package}{Character scalar, the name of the source package}
}
\value{
\code{NULL}
}
\description{
Copy the files of a packaged \code{drake} workflow into a supplied directory.
}
\examples{

\dontrun{

copy_pkg_files(directories = c("extdata","intdata"),
               overwrite_dirs = c("extdata" = TRUE,"intdata" = TRUE),
               package = "drakepkg")

}

}
