% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dir_links.R
\name{create_dir_links}
\alias{create_dir_links}
\title{Create the Links to the Inst/ Directory}
\usage{
create_dir_links(link_dir = "inst", targets = list())
}
\arguments{
\item{link_dir}{the path of the directory where the new links will be created (e.g., "inst")}

\item{targets}{a list of directory paths to which a link will be created}

\item{...}{parameters to be passed on to \code{R.utils::\link[R.utils]{createLink}()}}
}
\value{
\code{NULL}
}
\description{
Creates links between target directories and the \code{inst/} directory.
}
\examples{

\dontrun{

create_dir_links(link_dir = "inst"
               targets = list("extdata","intdata", "documents"),
               overwrite = TRUE)

}

}
