% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot.R
\name{create_plot}
\alias{create_plot}
\title{Create Plot}
\usage{
create_plot(data)
}
\arguments{
\item{data}{the \code{iris} dataset}
}
\value{
a \link{ggplot2} plot object
}
\description{
Create a specific histogram for the \code{iris} data
}
