% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce_analysis.R
\name{delete_analysis}
\alias{delete_analysis}
\title{Delete analysis directory structure}
\usage{
delete_analysis(to = getwd())
}
\arguments{
\item{to}{Character scalar, file path, where to delete the directories from
(default = current working directory).}
}
\description{
Delete the analysis directory structure of a packaged \code{drake} workflow from a
selected directory.
}
\note{
This function will delete a previously copied analysis directory structure from the user's
working directory (or the directory supplied to the \code{to} parameter).
}
