% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-plan.R
\name{example_plan}
\alias{example_plan}
\title{An Example of a Drake Plan}
\usage{
example_plan()
}
\value{
a \code{drake} plan object
}
\description{
An example of a drake plan that contains several targets
and commands, including processing input and output files.
}
\details{
The plan includes the following targets:
\itemize{
\item \code{raw_data}: \code{iris} data from an .xlsx file
\item \code{ready_data}: a prepared version of the \code{raw_data}
\item \code{hist}: a histogram plot created with \code{ggplot2}
\item \code{fit}: a linear model of the data (\code{Sepal.Width} as a function of \code{Petal.Width} and \code{Species})
}

These targets can be accessed using \code{drake::\link[drake]{loadd}} (e.g., \code{drake::loadd(fit)}).
}
