% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{external_data_list}
\alias{external_data_list}
\title{External Data List}
\format{The data itself is simply a \code{NULL} value.

Its purpose is to provide some documentation of the datasets that are not directly
included in the package (use \code{help(external_data_list)} to access it).

The following external datasets are accessed by workflows in this package:

\describe{

\item{\code{• example_one}:}{Brief description (\href{https://osf.io/u4kb7/}{link})}
\item{\code{• example_two}:}{Brief description (\href{https://osf.io/u4kb7/}{link})}
\item{\code{• example_three}:}{Brief description (\href{https://osf.io/u4kb7/}{link})}

}}
\usage{
external_data_list
}
\description{
This is a list of datasets that are included in the package workflow but
are not stored in the \code{data/} directory.

The primary reason that a dataset is included in this list is that it exceeds Github's
50 MB file size limit (see \href{https://help.github.com/articles/conditions-for-large-files/}{here} for details).
}
\keyword{datasets}
