% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\name{get_example_plan_external}
\alias{get_example_plan_external}
\title{An Example of a Drake Plan (External Data)}
\format{Returns a \code{\link[drake]{drake_plan}} object of class \code{tbl_df} with the following \code{\link[drake]{target}}s:
\describe{

\item{\code{raw_data}:}{\code{iris} data from an .xlsx file}
\item{\code{ready_data}:}{a prepared version of the \code{raw_data}}
\item{\code{hist}:}{a histogram plot created with \code{ggplot2}}
\item{\code{fit}:}{a linear model of the data (\code{Sepal.Width} as a function of \code{Petal.Width} and \code{Species})}
\item{\code{report}:}{a html report document}

}}
\usage{
get_example_plan_external()
}
\description{
An example of a \code{\link[drake]{drake_plan}} that includes data from external sources
(i.e., files not included in the pacakge).
}
\examples{

# Print the plan

print(get_example_plan())


# Make the plan, load a target, print the target

\dontrun{

make(get_example_plan_external())

loadd(fit)

print(fit)
}
}
\keyword{drake,}
\keyword{plan}
