% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduce_analysis.R
\name{get_file_list}
\alias{get_file_list}
\title{get file list}
\usage{
get_file_list(dir)
}
\arguments{
\item{dir}{Character, file path, where the files are searched for.}
}
\description{
get list of all files and directories in the selected directory that
are part of the analysis directory structure.
}
