% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_analysis.R
\name{get_plan}
\alias{get_plan}
\title{Get the analysis plan}
\usage{
get_plan()
}
\value{
\code{drake} workflow for the analysis
}
\description{
Generate the analysis pipeline within the \code{drake} framework.
}
