% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-iris-combined.R
\name{make_iris_combined}
\alias{make_iris_combined}
\title{Make Two Versions of Iris}
\usage{
make_iris_combined(iris_int_raw, iris_ext_raw)
}
\arguments{
\item{iris_int_raw}{the \code{iris} dataset stored within the \code{link{drakepkg}} package}

\item{iris_ext_raw}{the \code{iris} dataset downloaded from the web}
}
\value{
a \code{tibble} object
}
\description{
Make an  \code{tbl} object that combines the true \code{iris} dataset
with a resampled version.
}
