% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-iris-external.R
\name{make_iris_external}
\alias{make_iris_external}
\title{Make An Iris-like Dataset}
\usage{
make_iris_external()
}
\value{
a \code{tibble} object
}
\description{
Make an  \code{tbl} object that resembles the \code{iris} dataset.
THis object is not included in the package; instead, it is downloaded from
an external source and saved in the \code{extdata/} directory.
}
