% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_with_beep}
\alias{make_with_beep}
\title{Make a Beeping Plan!}
\usage{
make_with_beep(plan, .beep = TRUE, .beep_on_error = TRUE, ...)
}
\arguments{
\item{plan}{drake plan}

\item{.beep}{logical (default = TRUE), whether to play the "ping" sound after the plan is made}

\item{.beep_on_error}{logical (default = TRUE), whether to play the "facebook" sound if the plan returns an error}

\item{...}{list, a list of parameters to pass to \code{\link[drake]{make}()}}
}
\value{
nothing
}
\description{
Add a sound effect to alert the user when a \code{drake} plan has either completed or returned an error.
}
