% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf-check-write-access.R
\name{osf_check_write_access}
\alias{osf_check_write_access}
\title{Check OSF Write Access}
\usage{
osf_check_write_access(osf_id)
}
\arguments{
\item{osf_id}{character, the OSF project's unique identifier (e.g.,"e81xl")}
}
\value{
\code{TRUE} if the current OSF user has write access to the project, otherwise \code{FALSE}
}
\description{
Check OSF Write Access
}
\examples{
\dontrun{
  osf_check_write_access("tskpu","iris.csv", overwrite = TRUE)
  }
}
