% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf-get-file-version.R
\name{osf_get_file_version}
\alias{osf_get_file_version}
\title{Get OSF File Version Number}
\usage{
osf_get_file_version(osf_id = NULL)
}
\arguments{
\item{osf_id}{character, the OSF file's unique identifier (e.g., "tskpu")
to get the file info for}
}
\value{
a double

the file's version number (integer)
}
\description{
Get the version number of a file that is stored in a OSF project.
}
