% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf-upload-or-update.R
\name{osf_upload_or_update}
\alias{osf_upload_or_update}
\title{Upload/Update A File On OSF}
\usage{
osf_upload_or_update(has_osf_access = FALSE, project_id = NULL,
  file_id = NULL, path = NULL, osf_dirpath = NULL, name = NULL)
}
\arguments{
\item{has_osf_access}{Logical, Whether the user has access to the project's OSF repository.}

\item{project_id}{Character, The id of the OSF project.}

\item{file_id}{Character, The id of the existing OSF file (if any).}

\item{path}{Character, The file path.}

\item{osf_dirpath}{Character, The directory path where the file will be located in the OSF project (e.g., "data/raw-data")}

\item{name}{Character, The name of the destination file on OSF (if NULL, basename(path) will be used).}
}
\value{
A timestamp of when the file was created/modified.
}
\examples{
\dontrun{
  osf_upload_or_update(has_osf_access = TRUE,
                   project_id = "u4kb7",
                   path = "iris.csv",
                   osf_dirpath = "data/processed-data")

   osf_upload_or_update(has_osf_access = TRUE,
                   project_id = "u4kb7",
                   path = "iris.csv",
                   file_id = "XXXXX",  # <- replace this with the file id
                   osf_dirpath = "data/processed-data")

}
}
