% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otheriris}
\alias{otheriris}
\title{Edgar Anderson's Iris Data}
\format{\code{\link[datasets]{iris}} is a data frame with 150 cases (rows) and the following five (5) variables (columns):

\describe{

\item{\code{Sepal.Length}:}{Length of the sepals}
\item{\code{Sepal.Width}:}{Width of the sepals}
\item{\code{Petal.Length}:}{Length of the petals}
\item{\code{Species}:}{Species name)}

}}
\source{
\url{https://osf.io/u4kb7/}
}
\usage{
otheriris
}
\description{
This famous (Fisher's or Anderson's) iris data set gives the measurements in centimeters of the variables sepal length and width and petal length and width, respectively, for 50 flowers from each of 3 species of iris. The species are Iris setosa, versicolor, and virginica.
}
\keyword{datasets}
