% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-report.R
\name{write_html_report}
\alias{write_html_report}
\title{Write Report}
\usage{
write_html_report(hist, fit, rmd_filepath, html_filepath)
}
\arguments{
\item{hist}{a plot object}

\item{fit}{a linear model}

\item{rmd_filepath}{the filepath of the .Rmd version of the report}

\item{html_filepath}{the filepath of the .html version of the report}
}
\value{
a \code{tibble} object
}
\description{
Write the simple version of the report (.Rmd -> html)
}
