% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plans.R
\docType{data}
\name{plan_example}
\alias{plan_example}
\title{An Example of a Drake Plan}
\format{A \code{drake::\link[drake]{drake_plan}} object (\code{class=data.frame}) with the following targets:
\describe{

\item{\code{raw_data}:}{\code{iris} data from an .xlsx file}
\item{\code{ready_data}:}{ a prepared version of the \code{raw_data}}
\item{\code{hist}:}{ a histogram plot created with \code{ggplot2}}
\item{\code{fit}:}{ a linear model of the data (\code{Sepal.Width} as a function of \code{Petal.Width} and \code{Species})}
\item{\code{report}}{a html report document}

These targets can be accessed using \code{drake::\link[drake]{loadd}}
(e.g., \code{drake::loadd(fit)}).
}}
\usage{
plan_example
}
\description{
An example of a drake plan that contains several targets
and commands, including processing input and output files.
}
\examples{
  drake::make(plan_example)
}
\keyword{drake,}
\keyword{plan}
