% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-report.R
\name{write_report_simple}
\alias{write_report_simple}
\alias{write_report_external}
\title{Write Report}
\usage{
write_report_simple(hist, fit)

write_report_external(hist, fit)
}
\arguments{
\item{hist}{a plot object}

\item{fit}{a linear model}
}
\value{
a \code{tibble} object
}
\description{
Write the simple version of the report (.Rmd -> html)
}
